+++ api/Program.cs
@@ -79,6 +79,7 @@ HolidayRoutes.RegisterHolidayAPI(app);
 TypicalDishRoutes.RegisterTypicalDishAPI(app);
 TraditionalFairAndFestivalRoutes.RegisterTraditionalFairAndFestivalAPI(app);
 IntangibleHeritageRoutes.RegisterIntangibleHeritageAPI(app);
+HeritageCityRoutes.RegisterHeritageCityAPI(app);
 
 app.UseStatusCodePages(static context => {
     var request = context.HttpContext.Request;